requirements.txt
absl-py==0.11.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cloudpickle==1.6.0
Deprecated==1.2.12
docstring-parser==0.10
fire==0.4.0
google-api-core==1.31.1
google-api-python-client==1.12.8
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
jsonschema==3.2.0
kfp==1.7.0
kfp-pipeline-spec==0.1.9
kfp-server-api==1.6.0
kubernetes==12.0.1
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-poppler==0.2.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
strip-hints==0.1.10
tabulate==0.8.9
termcolor==1.1.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
websocket-client==1.2.0
wrapt==1.12.1
